backend/requirements.txt
fastapi[all]==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
motor==3.3.2 
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 # bcrypt 해싱
openai==1.35.7
numpy==1.26.4
pymongo[srv]==4.7.2